// frontend/src/generated/model/toolDefinitionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Polyglot V2 API
 * OpenAPI spec version: 0.1.0
 */
import type { ToolDefinitionResponseInputSchema } from "./toolDefinitionResponseInputSchema";
import type { ToolDefinitionResponseOutputSchema } from "./toolDefinitionResponseOutputSchema";

export interface ToolDefinitionResponse {
  effect: string;
  input_schema: ToolDefinitionResponseInputSchema;
  max_input_bytes: number;
  max_output_bytes: number;
  name: string;
  output_schema: ToolDefinitionResponseOutputSchema;
  roles: string[];
  timeout_ms: number;
  version: string;
}